+++ src/client.ts
@@ -1,8 +1,8 @@
 /// <reference path="types/libs.ts" />
 
 import { EventEmitter } from 'events';
-import StreamingClient from 'genesys-cloud-streaming-client';
 import StrictEventEmitter from 'strict-event-emitter-types';
+import StreamingClient, { HttpClient } from 'genesys-cloud-streaming-client';
 
 import {
   ISdkConfig,
@@ -15,13 +15,15 @@ import {
   IMediaDeviceIds,
   IUpdateOutgoingMedia,
   SdkEvents,
-  IMediaSettings
+  IMediaSettings,
+  isSecurityCode,
+  isCustomerData
 } from './types/interfaces';
 import {
   setupStreamingClient,
   proxyStreamingClientEvents
 } from './client-private';
-import { requestApi, createAndEmitSdkError, defaultConfigOption } from './utils';
+import { requestApi, createAndEmitSdkError, defaultConfigOption, requestApiWithRetry } from './utils';
 import { setupLogging } from './logging';
 import { SdkErrorTypes, SessionTypes } from './types/enums';
 import { SessionManager } from './sessions/session-manager';
@@ -77,12 +79,12 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
 
   _connected: boolean;
   _streamingConnection: StreamingClient;
+  _http: HttpClient;
   _orgDetails: any;
   _personDetails: IPersonDetails;
   _clientId: string;
   _customerData: ICustomerData;
   _hasConnected: boolean;
-  _refreshIceServersInterval: NodeJS.Timeout;
   _config: ISdkConfig;
 
   get isInitialized (): boolean {
@@ -144,9 +146,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
     // Telemetry for specific events
     // onPendingSession, onSession, onMediaStarted, onSessionTerminated logged in event handlers
     this.on('disconnected', this.logger.error.bind(this.logger, 'onDisconnected'));
-    this.on('cancelPendingSession', this.logger.info.bind(this.logger, 'cancelPendingSession'));
-    this.on('handledPendingSession', this.logger.info.bind(this.logger, 'handledPendingSession'));
-    this.on('trace', this.logger.info.bind(this.logger));
+    this.on('cancelPendingSession', (sessionId: string) => this.logger.info('cancelPendingSession', { sessionId }));
+    this.on('handledPendingSession', (sessionId: string) => this.logger.info('handledPendingSession', { sessionId }));
     this.on('sdkError', (error) => {
       /* this prints it more readable in the console */
       this.logger.error('sdkError', {
@@ -154,11 +155,13 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
         message: error.message,
         type: error.type,
         details: error.details
-      });
+        /* logging errors is dangerous because they could contain PII */
+      }, true);
     });
 
     this._connected = false;
     this._streamingConnection = null;
+    this._http = new HttpClient();
   }
 
   /**
@@ -180,7 +183,7 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
       let guestPromise: Promise<void>;
 
       /* if there is a securityCode, fetch conversation details */
-      if (this.isSecurityCode(opts)) {
+      if (isSecurityCode(opts)) {
         this.logger.debug('Fetching conversation details via secuirty code', opts.securityCode);
         guestPromise = requestApi.call(this, '/conversations/codes', {
           method: 'post',
@@ -188,13 +191,13 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
             organizationId: this._orgDetails.id,
             addCommunicationCode: opts.securityCode
           },
-          auth: false
+          noAuthHeader: true
         }).then(({ body }) => {
           this._customerData = body;
         });
 
         /* if no securityCode, check for valid customerData */
-      } else if (this.isCustomerData(opts)) {
+      } else if (isCustomerData(opts)) {
         this.logger.debug('Using customerData passed into the initialize', opts);
         guestPromise = Promise.resolve().then(() => {
           this._customerData = opts;
@@ -205,16 +208,16 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
 
       httpRequests.push(guestPromise);
     } else {
-      const getOrg = requestApi.call(this, '/organizations/me')
+      const getOrg = requestApiWithRetry.call(this, '/organizations/me').promise
         .then(({ body }) => {
           this._orgDetails = body;
-          this.logger.debug('debug', 'Organization details', body);
+          this.logger.debug('Fetched organization details', body, true); // don't log PII
         });
 
-      const getPerson = requestApi.call(this, '/users/me')
+      const getPerson = requestApiWithRetry.call(this, '/users/me').promise
         .then(({ body }) => {
           this._personDetails = body;
-          this.logger.debug('debug', 'Person details', body);
+          this.logger.debug('Fetched person details', body, true); // don't log PII
         });
 
       httpRequests.push(getOrg);
@@ -529,6 +532,7 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * @returns a promise that fullfils once the web socket has disconnected
    */
   disconnect (): Promise<any> {
+    this._http.stopAllRetries();
     return this._streamingConnection.disconnect();
   }
 
@@ -537,6 +541,7 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * @returns a promise that fullfils once the web socket has reconnected
    */
   reconnect (): Promise<any> {
+    this._http.stopAllRetries();
     return this._streamingConnection.reconnect();
   }
 
@@ -563,32 +568,6 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
     this.removeAllListeners();
     this.media.destroy();
     await this.disconnect();
-  }
-
-  async _refreshIceServers () {
-    if (!this._streamingConnection.connected) {
-      this.logger.warn('Tried to refreshIceServers but streamingConnection is not connected');
-      return;
-    }
-
-    try {
-      const services = (await this._streamingConnection.webrtcSessions.refreshIceServers()) || [];
-
-      if (!services.length) {
-        this.logger.error(new Error('refreshIceServers yielded no results'));
-        return;
-      }
-
-      const stunServers = services.filter((service) => service.type === 'stun');
-      if (!stunServers.length) {
-        this.logger.info('No stun servers received, setting iceTransportPolicy to "relay"');
-        this._streamingConnection._webrtcSessions.config.iceTransportPolicy = 'relay';
-      }
-    } catch (err) {
-      const errorMessage = 'GenesysCloud SDK failed to update TURN credentials. The application should be restarted to ensure connectivity is maintained.';
-      this.logger.warn(errorMessage, err);
-      throw createAndEmitSdkError.call(this, SdkErrorTypes.generic, errorMessage, err);
-    }
   }
 
   /**
@@ -622,21 +601,6 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
       });
     });
   }
-
-  private isSecurityCode (data: { securityCode: string } | ICustomerData): data is { securityCode: string } {
-    data = data as { securityCode: string };
-    return !!(data && data.securityCode);
-  }
-
-  private isCustomerData (data: { securityCode: string } | ICustomerData): data is ICustomerData {
-    data = data as ICustomerData;
-    return !!(data
-      && data.conversation
-      && data.conversation.id
-      && data.sourceCommunicationId
-      && data.jwt);
-  }
-
 }
 
 export default GenesysCloudWebrtcSdk;