+++ test/test-pages/common/sdk-controller.js
@@ -18,6 +18,7 @@ async function initWebrtcSDK (environmentData, _conversationsApi, noAuth, withDe
     initOptions = { securityCode: document.getElementById('security-key').value };
     options.organizationId = document.getElementById('org-id').value;
     options.autoConnectSessions = true;
+    localStorage.setItem('sdk_org_id', options.organizationId);
   } else {
     const accessToken = utils.getAccessToken();
     if (!accessToken) {