+++ test/unit/sessions/base-session-handler.test.ts
@@ -175,6 +175,27 @@ describe('updateOutgoingMedia()', () => {
     /* audio only */
     await handler.updateOutgoingMedia(session as any, { videoDeviceId: undefined, audioDeviceId: null });
     expect(startMediaSpy).toBeCalledWith({ video: undefined, audio: null, session });
+    await handler.updateOutgoingMedia(session as any, { videoDeviceId: videoDeviceId, audioDeviceId: undefined });
+    expect(startMediaSpy).toBeCalledWith({ video: videoDeviceId, audio: undefined, session });
+    startMediaSpy.mockReset();
+    startMediaSpy.mockResolvedValue(stream as any);
+
+    /* audio only */
+    await handler.updateOutgoingMedia(session as any, { videoDeviceId: undefined, audioDeviceId: audioDeviceId });
+    expect(startMediaSpy).toBeCalledWith({ video: undefined, audio: audioDeviceId, session });
+
+  });
+
+  it('should not call startMedia if both video and audio device ids are undefined or false', async () => {
+    const session = new MockSession();
+    session._outboundStream = new MockStream();
+    const stream = new MockStream({ video: true, audio: true });
+
+    const startMediaSpy = jest.spyOn(mockSdk.media, 'startMedia').mockResolvedValue(stream as any);
+
+    /* video and audio with no IDs */
+    await handler.updateOutgoingMedia(session as any, { videoDeviceId: false, audioDeviceId: undefined });
+    expect(startMediaSpy).not.toBeCalled();
   });
 
   it('should skip any screenshare tracks on the session', async () => {